# internal/httpserver/respond/respond.go
package respond

import (
	"encoding/json"
	"net/http"
)

const (
	CodeValidation   = "validation_error"
	CodeUnauthorized = "unauthorized"
	CodeNotFound     = "not_found"
	CodeInternal     = "internal_error"
	CodeRateLimited  = "rate_limited"
)

// ErrorResponse is the standard JSON error envelope returned by API
// endpoints on failure: {"error":{"code":...,"message":...}}.
type ErrorResponse struct {
	Error struct {
		Code    string `json:"code" example:"validation_error"`
		Message string `json:"message" example:"invalid request"`
	} `json:"error"`
}

func JSON(w http.ResponseWriter, status int, payload any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(payload)
}

func Error(w http.ResponseWriter, status int, code, message string) {
	body := ErrorResponse{}
	body.Error.Code = code
	body.Error.Message = message
	JSON(w, status, body)
}